+++ ActionTest/BaseProject/BaseProject/src/Game/CPlayer.h
@@ -34,7 +34,7 @@ public:
 	void AttackEnd() override;
 
 	// ƒ_ƒ[ƒW‚ğó‚¯‚é
-	void TakeDamage(int damage, CObjectBase* causer) override;
+	void TakeDamage(float damage, CObjectBase* causer) override;
 
 	/// <summary>
 	/// Õ“Ëˆ—
